Removed broken "base" definition.
In vinyl "base" does not refer to the *basename* (i.e. the unqualified file name). Instead it is described as "typically where a glob starts". As it already defaults to "cwd" and we have no meaningful alternative, it should not be set. This solves an issue where paths would be messed up because vinyl-fs will try to write to a nonsensical relative path.